@@ -1,6 +1,4 @@
-﻿using Domain.Entities;
-
-namespace Application.Common.Interfaces
+﻿namespace Application.Common.Interfaces
 {
     public interface IApplicationDbContext
     {
