@@ -1,18 +1,32 @@
 ﻿namespace Application.Maintainer.Categories.Commands.ToggleCategory;
-public record ToggleCategoryCommand(Guid CategoryId) : IRequest;
+public record ToggleCategoryCommand(Guid CategoryId) : IRequest<ApiResponse>;
 
-public class ToggleSubCategoryCommandHandler(IApplicationDbContext context) : IRequestHandler<ToggleCategoryCommand>
+public class ToggleSubCategoryCommandHandler(IApplicationDbContext context, IApiResponseService responseService) : IRequestHandler<ToggleCategoryCommand, ApiResponse>
 {
     private readonly IApplicationDbContext _context = context;
+    private readonly IApiResponseService _responseService = responseService;
 
-    public async Task Handle(ToggleCategoryCommand request, CancellationToken cancellationToken)
+    public async Task<ApiResponse> Handle(ToggleCategoryCommand request, CancellationToken cancellationToken)
     {
-        var entity = await _context.Categories.FindAsync([request.CategoryId], cancellationToken);
+        try
+        {
+            var entity = await _context.Categories.FindAsync([request.CategoryId], cancellationToken);
 
-        Guard.Against.NotFound(request.CategoryId, entity);
+            Guard.Against.NotFound(entity, $"No existe categoria con la Id {request.CategoryId}");
 
-        entity.ToggleActive();
+            entity.ToggleActive();
 
-        await _context.SaveChangesAsync(cancellationToken);
+            await _context.SaveChangesAsync(cancellationToken);
+
+            return _responseService.Success("El estado de la categoria ha sido actualizado exitosamente");
+        }
+        catch (Common.Exceptions.NotFoundException ex)
+        {
+            return _responseService.Fail(ex.Message);
+        }
+        catch (Exception ex)
+        {
+            return _responseService.Fail(ex.Message);
+        }
     }
 }
