@@ -1,6 +1,4 @@
-﻿using Domain.Entities;
-
-namespace Application.Maintainer.Inventories.Queries.GetInventories;
+﻿namespace Application.Maintainer.Inventories.Queries.GetInventories;
 
 public class InventoryDto
 {
