@@ -1,6 +1,4 @@
-﻿using Domain.Entities;
-
-namespace Application.Maintainer.Manufacturers.Queries.GetManufacturerById;
+﻿namespace Application.Maintainer.Manufacturers.Queries.GetManufacturerById;
 public class ManufacturerVM
 {
     public Guid ManufacturerId { get; set; }
